requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg[binary]==3.3.3
pydantic==2.7.4
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
loguru==0.7.2
alembic==1.13.1